Move DomainSuggestionState interface to types file
