+++ packages/data-stores/src/domain-suggestions/types.ts
@@ -1,3 +1,5 @@
+import type { DataStatus } from './constants';
+
 export interface DomainSuggestionQuery {
 	/**
 	 * True to include .blog subdomain suggestions
@@ -185,3 +187,30 @@ export interface DomainAvailability {
 }
 
 export type TimestampMS = ReturnType< typeof Date.now >;
+
+export interface DomainSuggestionState {
+	/**
+	 * The state of the DomainSuggestions e.g. pending, failure etc
+	 */
+	state: DataStatus;
+
+	/**
+	 * Domain suggestion data typically returned from the API
+	 */
+	data: Record< string, DomainSuggestion[] | undefined >;
+
+	/**
+	 * Error message
+	 */
+	errorMessage: string | null;
+
+	/**
+	 * Timestamp from last updated attempt
+	 */
+	lastUpdated: TimestampMS;
+
+	/**
+	 * Pending timestamp
+	 */
+	pendingSince: TimestampMS | undefined;
+}
